feat: todo jwt 적용

+++ src/todo/todo.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, Inject, BadRequestException } from '@nestjs/common';
+import {
+  Injectable,
+  Inject,
+  BadRequestException,
+  NotFoundException,
+} from '@nestjs/common';
 import { Repository } from 'typeorm';
 import { CreateTodoRequestDto } from './dtos/create-todo.dto';
 import { TodoResponseDto } from './dtos/todo.dto';
@@ -12,35 +17,56 @@ export class TodoService {
     private todoRepository: Repository<TodoEntity>,
   ) {}
 
-  async findAll(): Promise<TodoResponseDto[]> {
-    return this.todoRepository.find();
+  async findAll(userId: string): Promise<TodoResponseDto[]> {
+    return this.todoRepository.find({
+      where: { user: { id: userId } },
+      relations: ['user'],
+    });
   }
 
-  async findOne(id: string): Promise<TodoResponseDto | null> {
-    return this.todoRepository.findOneBy({ id });
+  async findOne(userId: string, id: string): Promise<TodoResponseDto> {
+    const todo = await this.todoRepository.findOne({
+      where: { id, user: { id: userId } },
+      relations: ['user'],
+    });
+    if (!todo) {
+      throw new NotFoundException(`Todo with id ${id} not found`);
+    }
+    return todo;
   }
 
   async create(
+    userId: string,
     createTodoRequestDto: CreateTodoRequestDto,
   ): Promise<TodoResponseDto> {
     const existingTodo = await this.todoRepository.findOneBy({
       title: createTodoRequestDto.title,
+      user: { id: userId },
     });
     if (existingTodo) {
       throw new BadRequestException('Todo already exists');
     }
-    return this.todoRepository.save(createTodoRequestDto);
+    const newTodo = this.todoRepository.create({
+      title: createTodoRequestDto.title,
+      isCompleted: createTodoRequestDto.isCompleted,
+      user: { id: userId },
+    });
+    return this.todoRepository.save(newTodo);
   }
 
   async update(
+    userId: string,
     id: string,
     updateTodoRequestDto: UpdateTodoRequestDto,
   ): Promise<TodoResponseDto | null> {
-    await this.todoRepository.update(id, updateTodoRequestDto);
-    return this.findOne(id);
+    await this.todoRepository.update(
+      { id, user: { id: userId } },
+      updateTodoRequestDto,
+    );
+    return this.findOne(userId, id);
   }
 
-  async delete(id: string): Promise<void> {
-    await this.todoRepository.delete(id);
+  async delete(userId: string, id: string): Promise<void> {
+    await this.todoRepository.delete({ id, user: { id: userId } });
   }
 }
